+++ internal/domain/audit_log_service.go
@@ -1,0 +1,6 @@
+package domain
+
+type AuditLogService interface {
+	Send(log AuditLog)
+	StartConsume()
+}